Extract TBTC chain interfaces to separate file
