+++ pkg/extensions/tbtc/tbtc.go
@@ -14,43 +14,7 @@ import (
 	eth "github.com/keep-network/keep-ecdsa/pkg/chain"
 )
 
-var logger = log.Logger("app-tbtc")
-
-// Handle represents a chain handle extended with TBTC-specific capabilities.
-type Handle interface {
-	eth.Handle
-
-	Deposit
-	TBTCSystem
-}
-
-// Deposit is an interface that provides ability to interact
-// with Deposit contracts.
-type Deposit interface {
-	// KeepAddress returns the underlying keep address for the
-	// provided deposit.
-	KeepAddress(depositAddress string) (string, error)
-
-	// RetrieveSignerPubkey retrieves the signer public key for the
-	// provided deposit.
-	RetrieveSignerPubkey(depositAddress string) error
-}
-
-// TBTCSystem is an interface that provides ability to interact
-// with TBTCSystem contract.
-type TBTCSystem interface {
-	// OnDepositCreated installs a callback that is invoked when an
-	// on-chain notification of a new deposit creation is seen.
-	OnDepositCreated(
-		handler func(depositAddress string),
-	) (subscription.EventSubscription, error)
-
-	// OnDepositRegisteredPubkey installs a callback that is invoked when an
-	// on-chain notification of a deposit's pubkey registration is seen.
-	OnDepositRegisteredPubkey(
-		handler func(depositAddress string),
-	) (subscription.EventSubscription, error)
-}
+var logger = log.Logger("extensions-tbtc")
 
 // InitializeExtensions initializes extensions specific to the TBTC application.
 func InitializeExtensions(ctx context.Context, handle Handle) error {
